+++ example/dapp/lib/utils/crypto/helpers.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:walletconnect_flutter_v2_dapp/models/chain_metadata.dart';
 import 'package:walletconnect_flutter_v2_dapp/utils/crypto/chain_data.dart';
 import 'package:walletconnect_flutter_v2_dapp/utils/crypto/eip155.dart';
-import 'package:walletconnect_flutter_v2_dapp/utils/crypto/kadena.dart';
 import 'package:walletconnect_flutter_v2_dapp/utils/crypto/solana_data.dart';
 
 String getChainName(String chain) {
@@ -32,7 +31,7 @@ List<String> getChainMethods(ChainType value) {
   if (value == ChainType.solana) {
     return SolanaData.methods.values.toList();
   } else if (value == ChainType.kadena) {
-    return Kadena.methods.values.toList();
+    return EIP155.methods.values.toList(); //Kadena.methods.values.toList();
   } else {
     return EIP155.methods.values.toList();
   }
@@ -42,7 +41,7 @@ List<String> getChainEvents(ChainType value) {
   if (value == ChainType.solana) {
     return SolanaData.events.values.toList();
   } else if (value == ChainType.kadena) {
-    return Kadena.events.values.toList();
+    return EIP155.events.values.toList(); //Kadena.events.values.toList();
   } else {
     return EIP155.events.values.toList();
   }