+++ example/dapp/lib/utils/test_data.dart
@@ -1,7 +1,5 @@
 import 'dart:convert';
 
-import 'package:kadena_dart_sdk/kadena_dart_sdk.dart';
-
 const String testSignData = 'Test sign data';
 String testSignTypedData(String address) => jsonEncode(
       {
@@ -40,80 +38,80 @@ const typedData =
 
 /// KADENA ///
 
-SignRequest createSignRequest({
-  required String networkId,
-  required String signingPubKey,
-  required String sender,
-  String code = '"hello"',
-  Map<String, dynamic>? data,
-  List<DappCapp> caps = const [],
-  String chainId = '1',
-  int gasLimit = 2000,
-  double gasPrice = 1e-8,
-  int ttl = 600,
-}) =>
-    SignRequest(
-      code: code,
-      data: data ?? {},
-      sender: sender,
-      networkId: networkId,
-      chainId: chainId,
-      gasLimit: gasLimit,
-      gasPrice: gasPrice,
-      signingPubKey: signingPubKey,
-      ttl: ttl,
-      caps: caps,
-    );
+// SignRequest createSignRequest({
+//   required String networkId,
+//   required String signingPubKey,
+//   required String sender,
+//   String code = '"hello"',
+//   Map<String, dynamic>? data,
+//   List<DappCapp> caps = const [],
+//   String chainId = '1',
+//   int gasLimit = 2000,
+//   double gasPrice = 1e-8,
+//   int ttl = 600,
+// }) =>
+//     SignRequest(
+//       code: code,
+//       data: data ?? {},
+//       sender: sender,
+//       networkId: networkId,
+//       chainId: chainId,
+//       gasLimit: gasLimit,
+//       gasPrice: gasPrice,
+//       signingPubKey: signingPubKey,
+//       ttl: ttl,
+//       caps: caps,
+//     );
 
-PactCommandPayload createPactCommandPayload({
-  required String networkId,
-  required String sender,
-  String code = '"hello"',
-  Map<String, dynamic>? data,
-  List<SignerCapabilities> signerCaps = const [],
-  String chainId = '1',
-  int gasLimit = 2000,
-  double gasPrice = 1e-8,
-  int ttl = 600,
-}) =>
-    PactCommandPayload(
-      networkId: networkId,
-      payload: CommandPayload(
-        exec: ExecMessage(
-          code: code,
-          data: data ?? {},
-        ),
-      ),
-      signers: signerCaps,
-      meta: CommandMetadata(
-        chainId: chainId,
-        gasLimit: gasLimit,
-        gasPrice: gasPrice,
-        ttl: ttl,
-        sender: sender,
-      ),
-    );
+// PactCommandPayload createPactCommandPayload({
+//   required String networkId,
+//   required String sender,
+//   String code = '"hello"',
+//   Map<String, dynamic>? data,
+//   List<SignerCapabilities> signerCaps = const [],
+//   String chainId = '1',
+//   int gasLimit = 2000,
+//   double gasPrice = 1e-8,
+//   int ttl = 600,
+// }) =>
+//     PactCommandPayload(
+//       networkId: networkId,
+//       payload: CommandPayload(
+//         exec: ExecMessage(
+//           code: code,
+//           data: data ?? {},
+//         ),
+//       ),
+//       signers: signerCaps,
+//       meta: CommandMetadata(
+//         chainId: chainId,
+//         gasLimit: gasLimit,
+//         gasPrice: gasPrice,
+//         ttl: ttl,
+//         sender: sender,
+//       ),
+//     );
 
-QuicksignRequest createQuicksignRequest({
-  required String cmd,
-  List<QuicksignSigner> sigs = const [],
-}) =>
-    QuicksignRequest(
-      commandSigDatas: [
-        CommandSigData(
-          cmd: cmd,
-          sigs: sigs,
-        ),
-      ],
-    );
+// QuicksignRequest createQuicksignRequest({
+//   required String cmd,
+//   List<QuicksignSigner> sigs = const [],
+// }) =>
+//     QuicksignRequest(
+//       commandSigDatas: [
+//         CommandSigData(
+//           cmd: cmd,
+//           sigs: sigs,
+//         ),
+//       ],
+//     );
 
-GetAccountsRequest createGetAccountsRequest({
-  required String account,
-}) =>
-    GetAccountsRequest(
-      accounts: [
-        AccountRequest(
-          account: account,
-        ),
-      ],
-    );
+// GetAccountsRequest createGetAccountsRequest({
+//   required String account,
+// }) =>
+//     GetAccountsRequest(
+//       accounts: [
+//         AccountRequest(
+//           account: account,
+//         ),
+//       ],
+//     );